refactor: Optimize image loading in blog page
This commit refactors the blog page component to improve image loading performance. It introduces the use of the `getPlaiceholder` library to generate a blurred placeholder image while the actual image is being fetched. This improves the perceived loading speed of the page and provides a better user experience.

The changes include:
- Importing the `getPlaiceholder` function from the `plaiceholder` library
- Fetching the blurred image using the `fetch` API and converting it to a `Buffer`
- Generating the base64-encoded blurred image using the `getPlaiceholder` function
- Adding the `placeholder` and `blurDataURL` attributes to the `ImageWithSuspense` component to display the blurred image while the actual image is loading

These changes optimize the image loading process in the blog page, resulting in a faster and smoother user experience.